feat: computer 분리
